+++ CyborgBattle/glob.h
@@ -31,7 +31,7 @@ public:
 	void updateDamage();
 
 	static void reset();
-	static float score();
+	static int score();
 private:
 	static int killed;  // Keep track of N of killed Globs
 	int state = GlobState::Idle;
